+++ Modules/Microsoft365DSC/DSCResources/MSFT_SCDLPComplianceRule/MSFT_SCDLPComplianceRule.psm1
@@ -51,7 +51,7 @@ function Get-TargetResource
         $GenerateIncidentReport,
 
         [Parameter()]
-        [ValidateSet("All", "Default", "DetectionDetails", "Detections", "DocumentAuthor", "DocumentLastModifier", "MatchedItem", "OriginalContent", "RulesMatched", "Service", "Severity", "Title")]
+        [ValidateSet("All", "Default", "DetectionDetails", "Detections", "DocumentAuthor", "DocumentLastModifier", "MatchedItem", "OriginalContent", "RulesMatched", "Service", "Severity", "Title", "RetentionLabel", "SensitivityLabel")]
         [System.String[]]
         $IncidentReportContent,
 
@@ -81,6 +81,62 @@ function Get-TargetResource
         [ValidateSet("Ignore", "RetryThenBlock")]
         [System.String]
         $RuleErrorAction,
+
+        [Parameter()]
+        [System.String[]]
+        $AnyOfRecipientAddressContainsWords,
+
+        [Parameter()]
+        [System.String[]]
+        $AnyOfRecipientAddressMatchesPatterns,
+
+        [Parameter()]
+        [System.String[]]
+        $ContentExtensionMatchesWords,
+
+        [Parameter()]
+        [System.String[]]
+        $ExceptIfContentExtensionMatchesWords,
+
+        [Parameter()]
+        [System.Boolean]
+        $RemoveRMSTemplate,
+
+        [Parameter()]
+        [System.Boolean]
+        $StopPolicyProcessing,
+
+        [Parameter()]
+        [System.Boolean]
+        $DocumentIsUnsupported,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExceptIfDocumentIsUnsupported,
+
+        [Parameter()]
+        [System.Boolean]
+        $HasSenderOverride,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExceptIfHasSenderOverride,
+
+        [Parameter()]
+        [System.Boolean]
+        $ProcessingLimitExceeded,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExceptIfProcessingLimitExceeded,
+
+        [Parameter()]
+        [System.Boolean]
+        $DocumentIsPasswordProtected,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExceptIfDocumentIsPasswordProtected,
 
         [Parameter()]
         [ValidateSet('Present', 'Absent')]
@@ -145,45 +201,68 @@ function Get-TargetResource
                 $NotifyAllowOverrideValue = $PolicyRule.NotifyAllowOverride.Replace(' ', '').Split(',')
             }
 
-            [array] $SensitiveInfoTypes = @()
-
-            foreach ($SensitiveInfo in $PolicyRule.ContentContainsSensitiveInformation)
+            if ($null -ne $PolicyRule.AnyOfRecipientAddressContainsWords -and $PolicyRule.AnyOfRecipientAddressContainsWords.count -gt 0)
             {
-                if ($null -ne $SensitiveInfo.groups)
-                {
-                    $groups = $SensitiveInfo.groups
-                    $SensitiveInfo = @()
-                    foreach ($group in $groups)
-                    {
-                        foreach ($siEntry in $group.sensitivetypes)
-                        {
-                            $SensitiveInfo += [System.Collections.Hashtable]$siEntry
-                        }
-                    }
-                }
-                $SensitiveInfoTypes += $SensitiveInfo
+                $AnyOfRecipientAddressContainsWords = $PolicyRule.AnyOfRecipientAddressContainsWords.Replace(' ', '').Split(',')
+            }
+
+            if ($null -ne $PolicyRule.AnyOfRecipientAddressMatchesPatterns -and $PolicyRule.AnyOfRecipientAddressMatchesPatterns -gt 0)
+            {
+                $AnyOfRecipientAddressMatchesPatterns = $PolicyRule.AnyOfRecipientAddressMatchesPatterns.Replace(' ', '').Split(',')
+            }
+
+            if ($null -ne $PolicyRule.ContentExtensionMatchesWords -and $PolicyRule.ContentExtensionMatchesWords.count -gt 0)
+            {
+                $ContentExtensionMatchesWords = $PolicyRule.ContentExtensionMatchesWords.Replace(' ', '').Split(',')
+            }
+
+            if ($null -ne $PolicyRule.ExceptIfContentExtensionMatchesWords -and $PolicyRule.ExceptIfContentExtensionMatchesWords.count -gt 0)
+            {
+                $ExceptIfContentExtensionMatchesWords = $PolicyRule.ExceptIfContentExtensionMatchesWords.Replace(' ', '').Split(',')
+            }
+
+            [array] $SensitiveInfo = @()
+
+            foreach ($si in $PolicyRule.ContentContainsSensitiveInformation)
+            {
+
+                $SensitiveInfo += [System.Collections.Hashtable]$si
             }
 
             $result = @{
-                Ensure                              = 'Present'
-                Name                                = $PolicyRule.Name
-                Policy                              = $PolicyRule.ParentPolicyName
-                AccessScope                         = $PolicyRule.AccessScope
-                BlockAccess                         = $PolicyRule.BlockAccess
-                BlockAccessScope                    = $PolicyRule.BlockAccessScope
-                Comment                             = $PolicyRule.Comment
-                ContentContainsSensitiveInformation = $SensitiveInfoTypes
-                ContentPropertyContainsWords        = $PolicyRule.ContentPropertyContainsWords
-                Disabled                            = $PolicyRule.Disabled
-                GenerateAlert                       = $PolicyRule.GenerateAlert
-                GenerateIncidentReport              = $PolicyRule.GenerateIncidentReport
-                IncidentReportContent               = $ArrayIncidentReportContent
-                NotifyAllowOverride                 = $NotifyAllowOverrideValue
-                NotifyEmailCustomText               = $PolicyRule.NotifyEmailCustomText
-                NotifyPolicyTipCustomText           = $PolicyRule.NotifyPolicyTipCustomText
-                NotifyUser                          = $PolicyRule.NotifyUser
-                ReportSeverityLevel                 = $PolicyRule.ReportSeverityLevel
-                RuleErrorAction                     = $PolicyRule.RuleErrorAction
+                Ensure                               = 'Present'
+                Name                                 = $PolicyRule.Name
+                Policy                               = $PolicyRule.ParentPolicyName
+                AccessScope                          = $PolicyRule.AccessScope
+                BlockAccess                          = $PolicyRule.BlockAccess
+                BlockAccessScope                     = $PolicyRule.BlockAccessScope
+                Comment                              = $PolicyRule.Comment
+                ContentContainsSensitiveInformation  = $SensitiveInfo
+                ContentPropertyContainsWords         = $PolicyRule.ContentPropertyContainsWords
+                Disabled                             = $PolicyRule.Disabled
+                GenerateAlert                        = $PolicyRule.GenerateAlert
+                GenerateIncidentReport               = $PolicyRule.GenerateIncidentReport
+                IncidentReportContent                = $ArrayIncidentReportContent
+                NotifyAllowOverride                  = $NotifyAllowOverrideValue
+                NotifyEmailCustomText                = $PolicyRule.NotifyEmailCustomText
+                NotifyPolicyTipCustomText            = $PolicyRule.NotifyPolicyTipCustomText
+                NotifyUser                           = $PolicyRule.NotifyUser
+                ReportSeverityLevel                  = $PolicyRule.ReportSeverityLevel
+                RuleErrorAction                      = $PolicyRule.RuleErrorAction
+                RemoveRMSTemplate                    = $PolicyRule.RemoveRMSTemplate
+                StopPolicyProcessing                 = $PolicyRule.StopPolicyProcessing
+                DocumentIsUnsupported                = $PolicyRule.DocumentIsUnsupported
+                ExceptIfDocumentIsUnsupported        = $PolicyRule.ExceptIfDocumentIsUnsupported
+                HasSenderOverride                    = $PolicyRule.HasSenderOverride
+                ExceptIfHasSenderOverride            = $PolicyRule.ExceptIfHasSenderOverride
+                ProcessingLimitExceeded              = $PolicyRule.ProcessingLimitExceeded
+                ExceptIfProcessingLimitExceeded      = $PolicyRule.ExceptIfProcessingLimitExceeded
+                DocumentIsPasswordProtected          = $PolicyRule.DocumentIsPasswordProtected
+                ExceptIfDocumentIsPasswordProtected  = $PolicyRule.ExceptIfDocumentIsPasswordProtected
+                AnyOfRecipientAddressContainsWords   = $AnyOfRecipientAddressContainsWords
+                AnyOfRecipientAddressMatchesPatterns = $AnyOfRecipientAddressMatchesPatterns
+                ContentExtensionMatchesWords         = $ContentExtensionMatchesWords
+                ExceptIfContentExtensionMatchesWords = $ExceptIfContentExtensionMatchesWords
             }
 
             $paramsToRemove = @()
@@ -282,7 +361,7 @@ function Set-TargetResource
         $GenerateIncidentReport,
 
         [Parameter()]
-        [ValidateSet("All", "Default", "DetectionDetails", "Detections", "DocumentAuthor", "DocumentLastModifier", "MatchedItem", "OriginalContent", "RulesMatched", "Service", "Severity", "Title")]
+        [ValidateSet("All", "Default", "DetectionDetails", "Detections", "DocumentAuthor", "DocumentLastModifier", "MatchedItem", "OriginalContent", "RulesMatched", "Service", "Severity", "Title", "RetentionLabel", "SensitivityLabel")]
         [System.String[]]
         $IncidentReportContent,
 
@@ -312,6 +391,63 @@ function Set-TargetResource
         [ValidateSet("Ignore", "RetryThenBlock")]
         [System.String]
         $RuleErrorAction,
+
+
+        [Parameter()]
+        [System.String[]]
+        $AnyOfRecipientAddressContainsWords,
+
+        [Parameter()]
+        [System.String[]]
+        $AnyOfRecipientAddressMatchesPatterns,
+
+        [Parameter()]
+        [System.String[]]
+        $ContentExtensionMatchesWords,
+
+        [Parameter()]
+        [System.String[]]
+        $ExceptIfContentExtensionMatchesWords,
+
+        [Parameter()]
+        [System.Boolean]
+        $RemoveRMSTemplate,
+
+        [Parameter()]
+        [System.Boolean]
+        $StopPolicyProcessing,
+
+        [Parameter()]
+        [System.Boolean]
+        $DocumentIsUnsupported,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExceptIfDocumentIsUnsupported,
+
+        [Parameter()]
+        [System.Boolean]
+        $HasSenderOverride,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExceptIfHasSenderOverride,
+
+        [Parameter()]
+        [System.Boolean]
+        $ProcessingLimitExceeded,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExceptIfProcessingLimitExceeded,
+
+        [Parameter()]
+        [System.Boolean]
+        $DocumentIsPasswordProtected,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExceptIfDocumentIsPasswordProtected,
 
         [Parameter()]
         [ValidateSet('Present', 'Absent')]
@@ -438,7 +574,7 @@ function Test-TargetResource
         $GenerateIncidentReport,
 
         [Parameter()]
-        [ValidateSet("All", "Default", "DetectionDetails", "Detections", "DocumentAuthor", "DocumentLastModifier", "MatchedItem", "OriginalContent", "RulesMatched", "Service", "Severity", "Title")]
+        [ValidateSet("All", "Default", "DetectionDetails", "Detections", "DocumentAuthor", "DocumentLastModifier", "MatchedItem", "OriginalContent", "RulesMatched", "Service", "Severity", "Title", "RetentionLabel", "SensitivityLabel")]
         [System.String[]]
         $IncidentReportContent,
 
@@ -470,6 +606,62 @@ function Test-TargetResource
         $RuleErrorAction,
 
         [Parameter()]
+        [System.String[]]
+        $AnyOfRecipientAddressContainsWords,
+
+        [Parameter()]
+        [System.String[]]
+        $AnyOfRecipientAddressMatchesPatterns,
+
+        [Parameter()]
+        [System.String[]]
+        $ContentExtensionMatchesWords,
+
+        [Parameter()]
+        [System.String[]]
+        $ExceptIfContentExtensionMatchesWords,
+
+        [Parameter()]
+        [System.Boolean]
+        $RemoveRMSTemplate,
+
+        [Parameter()]
+        [System.Boolean]
+        $StopPolicyProcessing,
+
+        [Parameter()]
+        [System.Boolean]
+        $DocumentIsUnsupported,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExceptIfDocumentIsUnsupported,
+
+        [Parameter()]
+        [System.Boolean]
+        $HasSenderOverride,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExceptIfHasSenderOverride,
+
+        [Parameter()]
+        [System.Boolean]
+        $ProcessingLimitExceeded,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExceptIfProcessingLimitExceeded,
+
+        [Parameter()]
+        [System.Boolean]
+        $DocumentIsPasswordProtected,
+
+        [Parameter()]
+        [System.Boolean]
+        $ExceptIfDocumentIsPasswordProtected,
+
+        [Parameter()]
         [ValidateSet('Present', 'Absent')]
         [System.String]
         $Ensure = 'Present',
@@ -499,10 +691,16 @@ function Test-TargetResource
 
     #region Test Sensitive Information Type
     # For each Desired SIT check to see if there is an existing rule with the same name
-    foreach ($sit in $ContentContainsSensitiveInformation)
+    if ($null -ne $ValuesToCheck['ContentContainsSensitiveInformation'])
+    {
+        $contentSITS = Get-SCDLPSensitiveInformation -SensitiveInformation $ValuesToCheck['ContentContainsSensitiveInformation']
+    }
+
+    foreach ($sit in $contentSITS)
     {
         Write-Verbose -Message "Trying to find existing Sensitive Information Action matching name {$($sit.name)}"
-        $matchingExistingRule = $CurrentValues.ContentContainsSensitiveInformaton | Where-Object -FilterScript { $_.name -eq $sit.name }
+        $matchingExistingRule = $CurrentValues.ContentContainsSensitiveInformation | Where-Object -FilterScript { $_.name -eq $sit.name }
+
         if ($null -ne $matchingExistingRule)
         {
             Write-Verbose -Message "Sensitive Information Action {$($sit.name)} was found"
@@ -534,28 +732,14 @@ function Test-TargetResource
         }
     }
     #endregion
+    $ValuesToCheck.Remove('ContentContainsSensitiveInformation') | Out-Null
+
+    Write-Verbose "Completed SIT check"
 
     $TestResult = Test-M365DSCParameterState -CurrentValues $CurrentValues `
         -Source $($MyInvocation.MyCommand.Source) `
         -DesiredValues $PSBoundParameters `
-        -ValuesToCheck @("Name",
-        "Policy",
-        "AccessScope",
-        "BlockAccess",
-        "BlockAccessScope",
-        "Comment",
-        "ContentPropertyContainsWords",
-        "Disabled",
-        "GenerateAlert",
-        "GenerateIncidentReport",
-        "IncidentReportContent",
-        "NotifyAllowOverride",
-        "NotifyEmailCustomText",
-        "NotifyPolicyTipCustomText",
-        "NotifyUser",
-        "ReportSeverityLevel",
-        "RuleErrorAction",
-        "Ensure")
+        -ValuesToCheck $ValuesToCheck.Keys
 
     Write-Verbose -Message "Test-TargetResource returned $TestResult"
 
@@ -610,7 +794,10 @@ function Export-TargetResource
             {
                 $IsCIMArray = $true
             }
-            $Results.ContentContainsSensitiveInformation = ConvertTo-SCDLPSensitiveInformationString -InformationArray $Results.ContentContainsSensitiveInformation
+            if ($null -ne $Results.ContentContainsSensitiveInformation )
+            {
+                $Results.ContentContainsSensitiveInformation = ConvertTo-SCDLPSensitiveInformationString -InformationArray $Results.ContentContainsSensitiveInformation
+            }
 
             $Results = Update-M365DSCExportAuthenticationResults -ConnectionMode $ConnectionMode `
                 -Results $Results
@@ -619,8 +806,10 @@ function Export-TargetResource
                 -ModulePath $PSScriptRoot `
                 -Results $Results `
                 -GlobalAdminAccount $GlobalAdminAccount
-            $currentDSCBlock = Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName "ContentContainsSensitiveInformation" -IsCIMArray $IsCIMArray
-
+            if ($null -ne $Results.ContentContainsSensitiveInformation )
+            {
+                $currentDSCBlock = Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName "ContentContainsSensitiveInformation" -IsCIMArray $IsCIMArray
+            }
             $dscContent += $currentDSCBlock
             Write-Host $Global:M365DSCEmojiGreenCheckMark
             $i++